Improved email-checking regex
Closer to the standard (see RFC 3696). For instance, the minimum size
email address is 3 (l@d) and most special characters (though not all)
are allowed in the local part.

Potential improvements involve allowing quoting characters that are not
usually allowed (\@), but, at that point, we'd probably just use a regex
written by someone else, such as the JavaMail optional library. And
since following the standard completely would also allow IP addresses in
square brackets as the domain, maybe that's not even what we want to do.

(also boolean zen)